refactor(process): split @gjsify/process/src/index.ts into per-concern modules
Follows the established pattern from @gjsify/webgl's webgl-context-base.ts
split (STATUS.md Workstream D). The 798-line `src/index.ts` becomes a
thin composer (74 LoC) on top of 8 per-concern modules.

**Layout:**

  src/internal/gjs.ts      — runtime-aware imports.gi / imports.system
                             access + Gio namespace probe
  src/internal/detect.ts   — version / platform / arch / pid / ppid
                             detection functions
  src/internal/env.ts      — process.env Proxy, argv, execPath, cwd, chdir
  src/internal/exit.ts     — process.exit with main-loop-aware scheduling
                             (GLib.idle_add hop so the syscall fires from a
                             fresh main-loop iteration; ensureMainLoop()
                             before quit so the source dispatches)
  src/internal/hrtime.ts   — hrtime (+ bigint) backed by GLib monotonic time
  src/internal/system.ts   — kill, memoryUsage, cpuUsage — all gated on
                             `isGjs()` so the native-fallback branch is
                             unreachable under GJS (where globalThis.process
                             IS this module, so unconditional delegation
                             would infinite-recurse)
  src/streams.ts           — ProcessReadStream (stdin) +
                             ProcessWriteStream (stdout/stderr)
  src/process-class.ts     — Process class wiring it all together
  src/index.ts             — instantiate + wire SIGWINCH + named exports

**`as any` reduction.** 17 → 3, all 3 now in `internal/gjs.ts` where
the `globalThis.imports` access pattern lives. The remaining `as any`s
are irreducible without typing the whole `imports.gi` namespace (a
moving target across GJS versions).

**Loop-safety guard generalized.** The original index.ts had multiple
`nativeProcess !== (this as any)` guards in instance methods to prevent
infinite recursion when `globalThis.process === <our Process instance>`.
Standalone helpers can't reference `this`, so the gate is now `isGjs()`
(checks `imports.gi.GLib`) — equivalent semantics, no `this` capture.

**Tests:** 148/148 Node + 148/148 GJS (unchanged — pure structural
refactor, no behavioural change). `yarn run check` clean.

**Why not split further?** The Process class itself stays at 125 lines;
splitting it would push state-management threading across files for
little gain. Keeping the singleton wiring in `process-class.ts` keeps
the SIGWINCH watcher + stream-construction order obvious.
